fix: correct real uint32 overflow and enum-check portability
The previous review-response commit (cdaa017) got two things wrong on
closer, independent inspection:

1. membrane_bench_packed_bytes's Q8_0 path computed
   (elems / MEMBRANE_QSIMD_BLOCK_ELEMS) * MEMBRANE_QSIMD_Q8_0_BLOCK_BYTES
   with `elems` left as uint32_t. Neither operand is wider than
   unsigned int, so C's usual arithmetic conversions evaluate this
   multiplication in 32-bit space -- and it genuinely wraps for elems
   near UINT32_MAX: (4294967264/32)*34 == 4563402718 truncates to
   268435422 in unsigned 32-bit arithmetic, before that result would
   ever reach the size_t return value. My prior comment claiming this
   was already safe was wrong -- it confused "the true 64-bit product
   fits in size_t" with "the C expression evaluates in size_t," which
   are different claims given the actual (narrower) operand types.
   Fixed by extracting membrane_bench_packed_bytes(uint32_t, int) and
   casting `elems` to size_t before the division, forcing the whole
   computation to happen at the right width; no separate SIZE_MAX guard
   is needed once that's true (elems/32*34's real maximum, ~4.56e9, is
   nowhere near overflowing a 64-bit size_t). Added a regression test
   exercising the exact wrap value above, without attempting the actual
   (4.5GB) allocation.

2. cfg_valid's enum bounds check relied on this compiler's specific
   choice of unsigned int as the representation for
   membrane_workload_kind_t/membrane_bench_policy_t -- correct on this
   toolchain, but not something C guarantees generally. Replaced with
   explicit (unsigned int) casts on both sides of each comparison, so
   the check is correct by construction rather than by this compiler's
   particular (if here confirmed) choice. Rebuilt with
   -Wall -Wextra -Wpedantic from a clean tree: zero warnings. The
   existing -1 regression tests (test_invalid_enum_values_rejected)
   still pass unmodified.

3. README's new Benchmark section previously showed concrete Storage%/
   MeanErr/MedianMs/Blocks/s example values. CONTRIBUTING.md requires
   "every claim in docs/ [and, per its later section, README.md] cite
   a source artifact ... checkable by scripts/verify-results.py" --
   nothing here is wired into that script, and building that wiring is
   out of this task's explicit scope (no new research-provenance
   hierarchy for a product benchmark). Replaced the concrete example
   row with a labeled column-shape description instead of real
   numbers: no fixed measurement remains in the permanent README.

Re-verified after these fixes: full clean rebuild is warning-free,
45/45 tests pass, membrane-demo's output is unchanged.

+++ tools/membrane-quant-policy-bench/bench_core.c
@@ -38,21 +38,23 @@ static double	median_of(double *samples, size_t n)
 	return ((samples[n / 2 - 1] + samples[n / 2]) / 2.0);
 }
 
-/* membrane_workload_kind_t/membrane_bench_policy_t have only
- * non-negative enumerators, so this compiler (verified: adding an
- * explicit lower-bound check here is flagged -Wtype-limits "always
- * true", i.e. the compiler itself confirms it) represents them as
- * unsigned int -- a bit pattern written via an out-of-range cast such
- * as (membrane_workload_kind_t)-1 reads back as a large positive value
- * here, which this single upper-bound check already rejects. */
+/* Explicit unsigned comparisons rather than relying on the enum's
+ * natural comparison: C leaves the underlying integer type of an enum
+ * with only non-negative enumerators implementation-defined, so an
+ * out-of-range value written via a cast like
+ * (membrane_workload_kind_t)-1 is only guaranteed to be rejected by
+ * `< COUNT` if every supported compiler picks an unsigned
+ * representation. Casting both sides to `unsigned int` makes the
+ * comparison correct regardless of that choice, without relying on it. */
 static int	cfg_valid(const membrane_bench_config_t *cfg)
 {
 	return (cfg->blocks != 0 && cfg->blocks <= MEMBRANE_BENCH_MAX_BLOCKS
 		&& cfg->iterations != 0
 		&& cfg->iterations <= MEMBRANE_BENCH_MAX_ITERATIONS
 		&& cfg->warmup <= MEMBRANE_BENCH_MAX_WARMUP
-		&& cfg->workload < MEMBRANE_WORKLOAD_COUNT
-		&& cfg->policy < MEMBRANE_BENCH_POLICY_COUNT);
+		&& (unsigned int)cfg->workload < (unsigned int)MEMBRANE_WORKLOAD_COUNT
+		&& (unsigned int)cfg->policy
+			< (unsigned int)MEMBRANE_BENCH_POLICY_COUNT);
 }
 
 /* One full pass over every pre-generated block, timed. Fills *acc (any
